Remove all workarounds for Go < 1.11
