+++ internal/backend/rclone/stdio_conn.go
@@ -5,6 +5,7 @@ import (
 	"os"
 	"os/exec"
 	"sync"
+	"time"
 
 	"github.com/restic/restic/internal/debug"
 )
@@ -56,6 +57,26 @@ func (s *StdioConn) LocalAddr() net.Addr {
 // RemoteAddr returns nil.
 func (s *StdioConn) RemoteAddr() net.Addr {
 	return Addr{}
+}
+
+// SetDeadline sets the read/write deadline.
+func (s *StdioConn) SetDeadline(t time.Time) error {
+	err1 := s.stdin.SetReadDeadline(t)
+	err2 := s.stdout.SetWriteDeadline(t)
+	if err1 != nil {
+		return err1
+	}
+	return err2
+}
+
+// SetReadDeadline sets the read/write deadline.
+func (s *StdioConn) SetReadDeadline(t time.Time) error {
+	return s.stdin.SetReadDeadline(t)
+}
+
+// SetWriteDeadline sets the read/write deadline.
+func (s *StdioConn) SetWriteDeadline(t time.Time) error {
+	return s.stdout.SetWriteDeadline(t)
 }
 
 // make sure StdioConn implements net.Conn
