--- internal/backend/rclone/stdio_conn_other.go
@@ -1,22 +1,0 @@
-// +build !go1.10
-
-package rclone
-
-import "time"
-
-// On Go < 1.10, it's not possible to set read/write deadlines on files, so we just ignore that.
-
-// SetDeadline sets the read/write deadline.
-func (s *StdioConn) SetDeadline(t time.Time) error {
-	return nil
-}
-
-// SetReadDeadline sets the read/write deadline.
-func (s *StdioConn) SetReadDeadline(t time.Time) error {
-	return nil
-}
-
-// SetWriteDeadline sets the read/write deadline.
-func (s *StdioConn) SetWriteDeadline(t time.Time) error {
-	return nil
-}
