+++ internal/test/helper.go
@@ -1,5 +1,3 @@
-// +build go1.9
-
 package test
 
 import "testing"
